+++ src/entropies_estimators/order_statistics/Vasicek.jl
@@ -73,5 +73,8 @@ function entropy(est::Vasicek, x::AbstractVector{T}) where {T<:Real}
         dprev = ith_order_statistic(ex, i - m, n)
         HVₘₙ += log(f * (dnext - dprev))
     end
-    return (HVₘₙ / n) / log(est.base, ℯ)
+
+    # The estimated entropy has "unit" [nats]
+    h = HVₘₙ / n
+    return convert_logunit(h, ℯ, est.base)
 end