+++ test/complexity/complexity.jl
@@ -3,4 +3,5 @@
     testfile("measures/missing_dispersion.jl")
     testfile("measures/entropy_approx.jl")
     testfile("measures/entropy_sample.jl")
+    testfile("measures/statistical_complexity.jl")
 end