+++ test/deprecations.jl
@@ -6,5 +6,5 @@ x = randn(1000)
 @test genentropy(x, 0.1) == entropy(Shannon(MathConstants.e), ValueHistogram(0.1), x)
 @test probabilities(x, 0.1) == probabilities(ValueHistogram(0.1), x)
 
-x = Dataset(rand(100, 3))
+x = StateSpaceSet(rand(100, 3))
 @test genentropy(x, 4) == entropy(Shannon(MathConstants.e), ValueHistogram(4), x)