+++ test/probabilities/estimators/transfer_operator.jl
@@ -1,15 +1,17 @@
-using StateSpaceSets: Dataset
-
+using ComplexityMeasures, Test
+using Random: MersenneTwister
 @test TransferOperator(RectangularBinning(3)) isa TransferOperator
 
-D = Dataset(rand(100, 2))
+D = StateSpaceSet(rand(MersenneTwister(1234), 100, 2))
 
+# Note that if `false` is used for `precise` the tests will fail.
+# But that's okay, since we do not do guarantees for that case.
 binnings = [
-    RectangularBinning(3),
-    RectangularBinning(0.2),
-    RectangularBinning([2, 3]),
-    RectangularBinning([0.2, 0.3]),
-    FixedRectangularBinning(range(0, 1; length = 5), 2)
+    RectangularBinning(3, true),
+    RectangularBinning(0.2, true),
+    RectangularBinning([2, 3], true),
+    RectangularBinning([0.2, 0.3], true),
+    FixedRectangularBinning(range(0, 1; length = 5), 2, true)
 ]
 
 # There's not easy way of constructing an analytical example for the resulting
@@ -20,7 +22,8 @@ binnings = [
 # each bin exactly.
 
 @testset "Binning test $i" for i in eachindex(binnings)
-    to = ComplexityMeasures.transferoperator(D, binnings[i])
+    b = binnings[i]
+    to = ComplexityMeasures.transferoperator(D, b)
     @test to isa ComplexityMeasures.TransferOperatorApproximationRectangular
 
     iv = invariantmeasure(to)
@@ -33,4 +36,7 @@ binnings = [
     est = TransferOperator(binnings[i])
     @test probabilities(est, D) isa Probabilities
     @test probabilities_and_outcomes(est, D) isa Tuple{Probabilities, Vector{SVector{2, Float64}}}
+
+    # Test that gives approximately same entropy as ValueHistogram:
+    abs(entropy(TransferOperator(b), D) - entropy(ValueHistogram(b), D) ) < 0.1 # or something like that
 end